pyproject.toml
[project]
name = "banking-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-jwt-extended>=4.7.1",
    "flask-sqlalchemy>=3.1.1",
    "logging>=0.4.9.6",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy-utils>=0.41.2",
    "uuid>=1.30",
]
